Move RHManageCategoryBase to management.py
